standard deviation approach to detecting "flat" response curves in OOD logit outputs
