--- methods/sigmoid_threshold.py
@@ -1,222 +1,0 @@
-#from __future__ import print_function
-
-import torch
-import torch.nn as nn
-import torch.optim as optim
-
-from torch.utils.data import DataLoader
-from utils.iterative_trainer import IterativeTrainerConfig, IterativeTrainer
-from utils.logger import Logger
-import os
-from os import path
-
-from methods.base_threshold import ProbabilityThreshold
-from datasets import MirroredDataset
-from torchinfo import summary
-
-class SigmoidThresholdWrapper(nn.Module):
-    """ The wrapper class for H.
-       
-    """
-    def __init__(self, base_model):
-        super(SigmoidThresholdWrapper, self).__init__()
-        self.base_model = base_model
-        output_size = base_model.output_size()[1].item()
-        """
-        self.H = nn.Sequential(
-                    nn.BatchNorm1d(output_size),
-                    nn.Linear(output_size, output_size)
-            )
-        self.sigmoid = nn.Sigmoid()
-        """
-        self.H = nn.Sigmoid()
-
-    def forward(self, x):
-        base_output = self.base_model.forward(x, softmax=False)
-        output = self.H(base_output)
-        #output = self.sigmoid(output)
-        return output
-    
-    def preferred_name(self):
-        return self.base_model.__class__.__name__
-    
-    def classify(self, x):
-        # we simply take the maximum. This is the basic prediction of the network
-        #print("x before:{}".format(x))
-        p = x.max(1)[1].unsqueeze_(1)
-        #print("x after:{}".format(p))
-        return p.long()
-
-    def get_output_device(self):
-        return self.base_model.get_output_device()
-
-    def get_info(self,args):
-        return self.base_model.get_info(args)
-
-class SigmoidThresholdClassifier(ProbabilityThreshold):
-    def method_identifier(self):
-        output = "SigmoidThresh"
-        if len(self.add_identifier) > 0:
-            output = output + "/" + self.add_identifier
-        return output
-
-    def propose_H(self, dataset):
-        config = self.get_base_config(dataset)
-
-        from models import get_ref_model_path
-        h_path = get_ref_model_path(self.args, config.model.__class__.__name__, dataset.name)
-        best_h_path = path.join(h_path, 'model.best.pth')
-
-        trainer = IterativeTrainer(config, self.args)
-
-        if not path.isfile(best_h_path):      
-            raise NotImplementedError("Please use model_setup to pretrain the networks first!")
-        else:
-            print('Loading H1 model from %s'%best_h_path)
-            config.model.load_state_dict(torch.load(best_h_path))
-        
-        trainer.run_epoch(0, phase='all')
-        test_average_acc = config.logger.get_measure('all_accuracy').mean_epoch(epoch=0)
-        print("All average accuracy %s"%(test_average_acc*100))
-
-        self.base_model = config.model
-        self.base_model.eval()
-
-    def get_H_config(self, dataset, will_train=True):
-        print("Preparing training D1+D2 (H)")
-        print("Mixture size:'%d'", len(dataset))
-        import global_vars as Global
-
-        # 80%, 20% for local train+test
-        train_ds, valid_ds = dataset.split_dataset(0.8)
-
-        if self.args.D1 in Global.mirror_augment:
-            print("Mirror augmenting %s"%self.args.D1)
-            new_train_ds = train_ds + MirroredDataset(train_ds)
-            train_ds = new_train_ds
-
-        # Initialize the multi-threaded loaders.
-        train_loader = DataLoader(train_ds, batch_size=self.args.batch_size, shuffle=True, num_workers=self.args.workers, pin_memory=True, drop_last=True)
-        valid_loader = DataLoader(valid_ds, batch_size=self.args.batch_size, num_workers=self.args.workers, pin_memory=True)
-        all_loader   = DataLoader(dataset,  batch_size=self.args.batch_size, num_workers=self.args.workers, pin_memory=True)
-
-        im,l = dataset[0]
-        input_size = im.size() #datasets in pytorch are assumed to be uniform size
-
-        # Set up the criterion
-        criterion = nn.CrossEntropyLoss().to(self.args.device)
-        criterion.size_average = True
-
-        # Set up the model
-        model = Global.get_ref_classifier(self.args.D1)[self.default_model]().to(self.args.device)
-        self.add_identifier = model.__class__.__name__
-        if hasattr(model, 'preferred_name'):
-            self.add_identifier = model.preferred_name()
-
-        #print("model before")
-        #summary(model, input_size=(self.args.batch_size, input_size[0], input_size[1], input_size[2]), depth=10)
-
-        model = SigmoidThresholdWrapper(model).to(self.args.device)
-
-        #print("model after")
-        #summary(model, input_size=(self.args.batch_size, input_size[0], input_size[1], input_size[2]), depth=10)
-
-        g = summary(model, input_size=(self.args.batch_size, input_size[0], input_size[1], input_size[2]), depth=100, verbose=0)
-        model_ram = g.to_megabytes(g.total_input) + g.float_to_megabytes(g.total_output + g.total_params)
-        print("Estimated Model Size:" + str(model_ram) + " Mb")
-
-        # Set up the config
-        config = IterativeTrainerConfig()
-
-        base_model_name = model.__class__.__name__
-        if hasattr(model, 'preferred_name'):
-            base_model_name = model.preferred_name()
-
-        config.name = '_%s[%s](%s-%s)'%(self.__class__.__name__, base_model_name, self.args.D1, self.args.D2)
-        config.train_loader = train_loader
-        config.valid_loader = valid_loader
-        config.phases = {
-                        'train':   {'dataset' : train_loader,  'backward': True},
-                        'test':    {'dataset' : valid_loader,  'backward': False},
-                        'testU':   {'dataset' : all_loader, 'backward': False},                        
-                        }
-        config.criterion = criterion
-        config.classification = True
-        config.cast_float_label = False
-        config.stochastic_gradient = True
-        config.visualize = not self.args.no_visualize
-        config.model = model
-        config.logger = Logger()
-        config.optim = optim.Adam(model.parameters(), lr=1e-3)
-        config.scheduler = optim.lr_scheduler.ReduceLROnPlateau(config.optim, patience=5, threshold=1e-2, min_lr=1e-6, factor=0.1, verbose=True)
-        config.max_epoch = 30
-        
-        if hasattr(model, 'train_config'):
-            model_train_config = model.train_config()
-            for key, value in model_train_config.items():
-                print('Overriding config.%s'%key)
-                config.__setattr__(key, value)
-
-        return config
-
-    def train_H(self, dataset):
-        # Wrap the (mixture)dataset in SubDataset so to easily
-        # split it later. God knows how many wrappers we have by this point.
-        from datasets import SubDataset
-        dataset = SubDataset('%s-%s'%(self.args.D1, self.args.D2), dataset, torch.arange(len(dataset)).int())
-
-        h_path = path.join(self.args.experiment_path, '%s'%(self.__class__.__name__),
-                                                      '%d'%(self.default_model),
-                                                      '%s-%s.pth'%(self.args.D1, self.args.D2))
-        h_parent = path.dirname(h_path)
-        if not path.isdir(h_parent):
-            os.makedirs(h_parent)
-
-        done_path = h_path + '.done'
-        will_train = self.args.force_train_h or not path.isfile(done_path)
-        
-        h_config = self.get_H_config(dataset)
-
-        trainer = IterativeTrainer(h_config, self.args)
-        """
-        if will_train:
-            print('Training from scratch')
-            best_accuracy = -1
-            trainer.run_epoch(0, phase='test')
-            for epoch in range(1, h_config.max_epoch):
-                trainer.run_epoch(epoch, phase='train')
-                trainer.run_epoch(epoch, phase='test')
-
-                train_loss = h_config.logger.get_measure('train_loss').mean_epoch()
-                h_config.scheduler.step(train_loss)
-
-                # Track the learning rates and threshold.
-                lrs = [float(param_group['lr']) for param_group in h_config.optim.param_groups]
-                h_config.logger.log('LRs', lrs, epoch)
-                h_config.logger.get_measure('LRs').legend = ['LR%d'%i for i in range(len(lrs))]
-
-                test_average_acc = h_config.logger.get_measure('test_accuracy').mean_epoch()
-
-                # Save the logger for future reference.
-                torch.save(h_config.logger.measures, path.join(h_parent, 'logger.%s-%s.pth'%(self.args.D1, self.args.D2)))
-
-                if best_accuracy < test_average_acc:
-                    print('Updating the on file model with %s'%('%.4f'%test_average_acc))
-                    best_accuracy = test_average_acc
-                    torch.save(h_config.model.state_dict(), h_path)
-
-                if test_average_acc > 1-1e-4:
-                    break
-
-            torch.save({'finished':True}, done_path)
-
-        # Load the best model.
-        print('Loading H model from %s'%h_path)
-        h_config.model.load_state_dict(torch.load(h_path))
-        """
-        trainer.run_epoch(0, phase='testU')
-        test_average_acc = h_config.logger.get_measure('testU_accuracy').mean_epoch(epoch=0)
-        print("Valid/Test average accuracy %s"%('%.4f%%'%(test_average_acc*100)))
-        self.H_class = h_config.model
-        self.H_class.eval()
-        return test_average_acc
